Bug with fixed by changing int to long Pell

+++ src/Main.java
@@ -31,8 +31,8 @@ public class Main {
 
 
     public static long pellNumbers(int n) {
-        int result = n;
-        for (int i = 2, a = 1, b = 2; i < n; i ++) {
+        long result = n;
+        for (long i = 2, a = 1, b = 2; i < n; i ++) {
             result = 2 * b + a;
             a = b;
             b = result;
